+++ cloudrun.tf
@@ -6,7 +6,7 @@ data "google_cloud_run_service" "cloud_run_service" {
 
 resource "google_compute_region_network_endpoint_group" "cloudrun_neg" {
   for_each              = var.services
-  name                  = "${each.key}-${local.random_suffix}"
+  name                  = var.use_random_postfix_for_network_endpoint_group ? "${each.key}-${local.random_suffix}" : each.key
   network_endpoint_type = "SERVERLESS"
   region                = lookup(each.value, "location", var.region)
   cloud_run {