+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v5.2.0] - 2022-08-21
+### Added
+- support for paths
+- random suffix size to easy migrations
+- add custom_url_map_name, custom_target_http_proxy_custom_url_map_name, use_random_postfix_for_network_endpoint_group, global_forwarding_rule_name to easy migrations
+
 ## [v5.1.0] - 2022-08-21
 ### Added
 - support for HTTPS type of backend